Extracted pizza class into it's own file
